writer/requirements.txt
nats-py
webcolors
pydantic
pytest